AMD_FH_WEBAPP/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
